split CSVMetaArray into reader and processor
